Fix unused-parameter and sign-compare warnings in kvfs/cfs
- kvfs/server/file.h, subtable.h: comment out unused session params in Make/Load stubs
- kvfs/server/publisher.cc: comment out unused next params in MakeFile/Unlink
- kvfs/server/fs.cc: comment out unused nblocks/block_size params in Make
- kvfs/client/client.cc: remove unused ret variable in Init(const char*)
- kvfs/client/table.cc: cast n to int to silence sign-compare warning
- cfs/server/fs.cc: remove unused childino variable in Unlink

+++ libfs/src/kvfs/client/table.cc
@@ -98,7 +98,7 @@ Table::Put(::client::Session* session, const char* key, const char* src, uint64_
 			goto done;
 		}
 	}
-	if ((ret = obj->Write(session, src, 0, n)) != n) {
+	if ((ret = obj->Write(session, src, 0, n)) != (int)n) {
 		ret = -E_NOMEM;
 		goto done;
 	}
@@ -260,7 +260,7 @@ Table::Put(::client::Session* session, const char* key, const char* src, uint64_
 		}
 	}
 #if 0
-	if ((ret = obj->Write(session, src, 0, n)) != n) {
+	if ((ret = obj->Write(session, src, 0, n)) != (int)n) {
 		ret = -E_NOMEM;
 		goto done;
 	}
